Refactor routes and templates, remove admin route, update middleware and config
