+++ src/middleware/no-auth.ts
@@ -27,7 +27,7 @@ export const requireNoAuth = async (c: Context, next: Next) => {
     await next();
 
     return;
-  }
+    }
 
-  return c.redirect('/admin');
+  return c.redirect('/dashboard');
 };
