+++ src/templates/login.tsx
@@ -8,7 +8,7 @@ export default function LoginPage() {
       <div class={styles.loginContainer}>
         <div class={styles.loginForm}>
           <h2 class={styles.loginFormH2}>🔐 Admin Login</h2>
-          <form id="loginForm" method="post" action="/admin/login">
+          <form id="loginForm" method="post" action="/auth/login">
             <div class={styles.formGroup}>
               <label for="username" class={styles.formGroupLabel}>
                 Username:
@@ -45,7 +45,7 @@ export default function LoginPage() {
           <p class="mt-4 text-center text-sm text-gray-600">
             Don't have an account?{' '}
             <a
-              href="/admin/register"
+              href="/auth/register"
               class="font-medium text-indigo-600 hover:text-indigo-500"
             >
               Create one here
